feat(mobile): circle-clip avatar viewer + banner viewer on profile cover photo
Follow-ups to #14444:

1. **Avatar viewer keeps the circle.** The avatar on the profile screen is
   a circle, so the full-screen viewer should clip to a circle too — not a
   letterboxed square on black. AvatarViewer now renders the image inside
   a fixed-size square wrapper with `borderRadius: size / 2` +
   `overflow: 'hidden'` and `resizeMode='cover'`, sized to fit the smaller
   window dimension minus 24px padding so it doesn't touch the edges.

2. **Banner viewer on the profile cover photo.** Tapping the cover photo
   at the top of the profile screen now opens a parallel full-screen
   viewer for the banner. Uses `WidthSizes.SIZE_2000` (the largest cached
   cover-photo size), `resizeMode='contain'` to preserve the cover photo's
   aspect ratio on the black backdrop. No circular clip.

To avoid duplicating the modal/gesture/close-button machinery, factored
the chrome out of the existing AvatarViewer into a shared
**FullscreenImageViewer** that owns:
- `<Modal transparent animationType='fade' statusBarTranslucent>`
- The pan gesture (Reanimated shared values, 120 px / 800 px-s dismiss
  thresholds, 200 ms spring-back, translation reset on open)
- The X close button (top-right, safe-area top inset + 8 px, hitSlop)
- The black `#000` backdrop + `StatusBar light-content`

AvatarViewer and BannerViewer are now thin wrappers around it that
describe their image shape and source. Same dismiss UX in both.

ProfileCoverPhoto wraps the existing `<CoverPhoto>` in a `<Pressable>`
that opens the BannerViewer; the artist-badge overlay (which has no own
press handler) sits absolute on top and stays visually unchanged.

Files:
- packages/mobile/src/screens/profile-screen/FullscreenImageViewer.tsx (new)
- packages/mobile/src/screens/profile-screen/AvatarViewer.tsx (refactor)
- packages/mobile/src/screens/profile-screen/BannerViewer.tsx (new)
- packages/mobile/src/screens/profile-screen/ProfileCoverPhoto.tsx (tap wiring)
