+++ packages/mobile/src/screens/profile-screen/ProfileCoverPhoto.tsx
@@ -1,6 +1,8 @@
+import { useCallback, useState } from 'react'
+
 import { useUserByParams } from '@audius/common/api'
 import { BlurView } from '@react-native-community/blur'
-import { StyleSheet } from 'react-native'
+import { Pressable, StyleSheet } from 'react-native'
 import { useCurrentTabScrollY } from 'react-native-collapsible-tab-view'
 import Animated, {
   interpolate,
@@ -13,6 +15,8 @@ import BadgeArtist from 'app/assets/images/badgeArtist.svg'
 import { CoverPhoto } from 'app/components/image/CoverPhoto'
 import { useRoute } from 'app/hooks/useRoute'
 import { makeStyles } from 'app/styles'
+
+import { BannerViewer } from './BannerViewer'
 
 const AnimatedBlurView = Animated.createAnimatedComponent(BlurView)
 
@@ -31,6 +35,10 @@ const useStyles = makeStyles(({ spacing }) => ({
     bottom: spacing(2)
   }
 }))
+
+const messages = {
+  viewBanner: 'View profile banner'
+}
 
 export const ProfileCoverPhoto = () => {
   const styles = useStyles()
@@ -52,6 +60,10 @@ export const ProfileCoverPhoto = () => {
 
   const isArtist = !!track_count && track_count > 0
   const coverPhotoHeight = insets.top + COVER_PHOTO_CONTENT_HEIGHT
+
+  const [isViewerOpen, setIsViewerOpen] = useState(false)
+  const handleOpenViewer = useCallback(() => setIsViewerOpen(true), [])
+  const handleCloseViewer = useCallback(() => setIsViewerOpen(false), [])
 
   const blurViewStyle = useAnimatedStyle(() => ({
     ...StyleSheet.absoluteFillObject,
@@ -81,19 +93,35 @@ export const ProfileCoverPhoto = () => {
     >
       {user_id ? (
         <>
-          <CoverPhoto style={{ height: coverPhotoHeight }} userId={user_id}>
-            <AnimatedBlurView
-              blurType='dark'
-              blurAmount={100}
-              style={blurViewStyle}
-            />
-            {isArtist ? <Animated.View style={styles.darkOverlay} /> : null}
-          </CoverPhoto>
+          {/* Pressable wraps the CoverPhoto so the banner area is tappable
+              to open the full-screen viewer. The artist badge (when
+              present) renders absolute on top — taps in the badge corner
+              still hit the Pressable underneath since the badge has no
+              own onPress handler. */}
+          <Pressable
+            onPress={handleOpenViewer}
+            accessibilityRole='imagebutton'
+            accessibilityLabel={messages.viewBanner}
+          >
+            <CoverPhoto style={{ height: coverPhotoHeight }} userId={user_id}>
+              <AnimatedBlurView
+                blurType='dark'
+                blurAmount={100}
+                style={blurViewStyle}
+              />
+              {isArtist ? <Animated.View style={styles.darkOverlay} /> : null}
+            </CoverPhoto>
+          </Pressable>
           {isArtist ? (
             <Animated.View style={[styles.artistBadge, badgeStyle]}>
               <BadgeArtist />
             </Animated.View>
           ) : null}
+          <BannerViewer
+            userId={user_id}
+            isOpen={isViewerOpen}
+            onClose={handleCloseViewer}
+          />
         </>
       ) : null}
     </Flex>
